Re-add DropdownSelector
